create a new module for comment count function
