+++ src/modules/commentCounter.js
@@ -1,0 +1,11 @@
+const countComment = (parentNode) => {
+  const number = parentNode.childElementCount-3
+  return number
+}
+
+const printComment = (number) => {
+  const commentTitle = document.getElementById('comment-title')
+  commentTitle.innerHTML = `Comments(${number})`
+}
+
+export {countComment, printComment}
